requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
motor==3.6.0
pydantic==2.10.2
pydantic-settings==2.6.1
openai==1.55.3
httpx==0.27.2
PyJWT==2.10.0
python-dotenv==1.0.1
sentry-sdk[fastapi]==2.19.2
pymongo==4.10.1
requests==2.32.3
python-multipart==0.0.12
bson
